soft delete comment only admin and user && update only admin and user

+++ routers/controllers/comment.js
@@ -26,18 +26,41 @@ const newComment = (req, res) => {
 const deleteCommet = (req, res) => {
   const { _id } = req.params;
   try {
-    commentModel
-      .findOneAndDelete({ _id: _id })
-      .then((result) => {
-        if (result) {
-          res.status(200).json(result);
+    commentModel.findOne({ _id: _id }).then((item) => {
+      if (item) {
+        if (item.user == req.token._id) {
+          commentModel
+            .findOneAndDelete({ _id: _id })
+            .then((result) => {
+              if (result) {
+                res.status(200).json(result);
+              } else {
+                res.status(404).send("Not found");
+              }
+            })
+            .catch((err) => {
+              res.status(400).json(err);
+            });
+        } else if (req.token.role == "61a734cd947e8eba47efbc68") {
+          commentModel
+            .findOneAndDelete({ _id: _id })
+            .then((result) => {
+              if (result) {
+                res.status(200).json(result);
+              } else {
+                res.status(404).send("Not found");
+              }
+            })
+            .catch((err) => {
+              res.status(400).json(err);
+            });
         } else {
-          res.status(404).send("Not found");
+          res.status(403).send("Forbidden");
         }
-      })
-      .catch((err) => {
-        res.status(400).json(err);
-      });
+      } else {
+        res.status(404).send("Not found");
+      }
+    });
   } catch (error) {
     res.status(400).json(error);
   }
@@ -48,20 +71,39 @@ const updateComment = (req, res) => {
   const { desc } = req.body;
   try {
     commentModel.findOne({ _id: _id }).then((item) => {
+      console.log(req.token);
+      if (item.user == req.token._id) {
+        commentModel
+          .findOneAndUpdate(
+            { _id: _id },
+            { $set: { desc: desc, time: Date() } },
+            { new: true }
+          )
+          .then((result) => {
+            if (result) {
+              res.status(200).json(result);
+            } else {
+              res.status(404).send("Comment not found");
+            }
+          });
+      } else if (req.token.role == "61a734cd947e8eba47efbc68") {
+        commentModel
+          .findOneAndUpdate(
+            { _id: _id },
+            { $set: { desc: desc, time: Date() } },
+            { new: true }
+          )
+          .then((result) => {
+            if (result) {
+              res.status(200).json(result);
+            } else {
+              res.status(404).send("Comment not found");
+            }
+          });
+      } else {
+        res.status(404).send("Forbidden");
+      }
     });
-    commentModel
-      .findOneAndUpdate(
-        { _id: _id },
-        { $set: { desc: desc, time: Date() } },
-        { new: true }
-      )
-      .then((result) => {
-        if (result) {
-          res.status(200).json(result);
-        } else {
-          res.status(404).send("Comment not found");
-        }
-      });
   } catch (error) {
     res.status(404).json(error);
   }
